requirements.txt
# --- Core API Stack ---
fastapi==0.121.1
uvicorn==0.38.0
sqlmodel==0.0.27
SQLAlchemy==2.0.44
aiomysql==0.2.0
pymysql==1.1.0
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0
python-multipart==0.0.5
python-dotenv==1.2.1
loguru==0.7.3
jinja2==3.1.6
slowapi==0.1.9
starlette==0.49.3
gunicorn==21.2.0

# --- Captcha Service ---
captcha==0.6.0

# --- PDF Generation (UPDATED: WeasyPrint) ---
weasyprint==63.1
Pillow==11.1.0
qrcode==8.0

# --- Database & Supabase ---
supabase==2.26.0
postgrest==2.26.0
storage3==2.26.0
realtime==2.26.0
alembic==1.17.1
greenlet==3.2.4

# --- Auth & Security ---
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
cryptography==46.0.3
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1

# --- Dependencies & Utilities ---
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0
anyio==4.11.0
cffi==2.0.0
pycparser==2.23
six==1.17.0
wrapt==2.0.1
deprecated==1.3.1
deprecation==2.1.0
tzdata==2025.2

redis==5.0.1
ddtrace==2.21.10